requirements.txt
crewai[litellm]
duckdb
python-dotenv
streamlit
pandas
